requirements.txt
# MinIO MCP Server Dependencies

# FastMCP for MCP server implementation
fastmcp>=0.8.0

# HTTP client for API calls
httpx>=0.24.0

# Environment variable management
python-dotenv>=1.0.0

# Data validation and settings
pydantic>=2.0.0

# JWT token handling
pyjwt>=2.6.0

# Additional utilities
cryptography>=3.4.8